+++ Classes/LZMASDK/7zIn.c
@@ -1477,7 +1477,7 @@ SzArEx_DictCache_mmap(SzArEx_DictCache *dictCache)
   
   // Make sure mapSize is in terms of whole pages
   {
-    int numPages = mapSize / SM_PAGESIZE;
+    int numPages = (int) mapSize / SM_PAGESIZE;
     if ((mapSize % SM_PAGESIZE) > 0) {
       numPages += 1;
     }
@@ -1502,7 +1502,7 @@ SzArEx_DictCache_mmap(SzArEx_DictCache *dictCache)
     // Need to actually write a byte in order for the file
     // to be extended to this length.
     char oneByte = 0;
-    int writeResult = fwrite(&oneByte, 1, 1, mapfile);
+    int writeResult = (int) fwrite(&oneByte, 1, 1, mapfile);
     assert(writeResult == 1);
     
     fflush(mapfile);
